+++ tests/phpstan/bootstrap.php
@@ -1,5 +1,5 @@
 <?php
 
-define( 'NOVAMIRA_MAINWP_VERSION', '0.7.0' );
+define( 'NOVAMIRA_MAINWP_VERSION', '0.7.1' );
 define( 'NOVAMIRA_MAINWP_FILE', __DIR__ . '/../../mainwp-novamira-addon.php' );
 define( 'NOVAMIRA_MAINWP_DIR', __DIR__ . '/../../' );